Added various guards in the main store and the annotation store

+++ src/store/index.ts
@@ -630,7 +630,9 @@ export class Main extends VuexModule {
 
   @Action
   async deleteLargeImage(largeImage: IGirderLargeImage) {
-    if (!this.dataset?.id || !largeImage._id) return;
+    if (!this.dataset?.id || !largeImage._id || !this.isLoggedIn) {
+      return;
+    }
 
     // Do not delete the default large image (original data)
     if (largeImage.name === DEFAULT_LARGE_IMAGE_SOURCE) {
@@ -805,6 +807,9 @@ export class Main extends VuexModule {
     datasetId: string;
     configurationId: string;
   }) {
+    if (!this.isLoggedIn) {
+      return null;
+    }
     return this.api.createDatasetView({
       datasetId,
       configurationId,
@@ -1072,6 +1077,9 @@ export class Main extends VuexModule {
 
   @Action
   deleteDatasetView(datasetView: IDatasetView) {
+    if (!this.isLoggedIn) {
+      return;
+    }
     return this.api.deleteDatasetView(datasetView.id);
   }
 
@@ -1085,6 +1093,9 @@ export class Main extends VuexModule {
     description: string;
     path: IGirderSelectAble;
   }) {
+    if (!this.isLoggedIn) {
+      return null;
+    }
     try {
       sync.setSaving(true);
       const ds = await this.api.createDataset(name, description, path);
@@ -1119,7 +1130,7 @@ export class Main extends VuexModule {
     description: string;
     folderId: string;
   }) {
-    if (!this.dataset) {
+    if (!this.dataset || !this.isLoggedIn) {
       return null;
     }
     try {
@@ -1218,6 +1229,9 @@ export class Main extends VuexModule {
 
   @Action
   async deleteConfiguration(configuration: IDatasetConfiguration) {
+    if (!this.isLoggedIn) {
+      return;
+    }
     try {
       sync.setSaving(true);
       const promises: Promise<any>[] = [];
@@ -1251,6 +1265,9 @@ export class Main extends VuexModule {
 
   @Action
   async deleteDataset(dataset: IDataset) {
+    if (!this.isLoggedIn) {
+      return;
+    }
     try {
       sync.setSaving(true);
       const promises: Promise<any>[] = [];
@@ -1299,7 +1316,9 @@ export class Main extends VuexModule {
       return;
     }
     this.setLastLocationInDatasetView(location);
-    await this.api.updateDatasetView(this.datasetView);
+    if (this.isLoggedIn) {
+      await this.api.updateDatasetView(this.datasetView);
+    }
   }
 
   @Action
@@ -1380,7 +1399,7 @@ export class Main extends VuexModule {
 
   @Action
   async syncConfiguration(key: keyof IDatasetConfigurationBase) {
-    if (!this.configuration) {
+    if (!this.configuration || !this.isLoggedIn) {
       return;
     }
     sync.setSaving(true);
@@ -1395,7 +1414,7 @@ export class Main extends VuexModule {
 
   @Action
   async renameConfiguration(newName: string) {
-    if (!this.configuration) {
+    if (!this.configuration || !this.isLoggedIn) {
       return;
     }
     sync.setSaving(true);
@@ -1410,7 +1429,7 @@ export class Main extends VuexModule {
 
   @Action
   async addLayer() {
-    if (!this.configuration || !this.dataset) {
+    if (!this.configuration || !this.dataset || !this.isLoggedIn) {
       return;
     }
     this.pushLayer(newLayer(this.dataset, this.layers, this.userChannelColors));
@@ -1598,7 +1617,9 @@ export class Main extends VuexModule {
     this.changeLayer({ layerId, delta: { contrast }, sync: true });
     if (this.datasetView) {
       Vue.delete(this.datasetView.layerContrasts, layerId);
-      this.api.updateDatasetView(this.datasetView);
+      if (this.isLoggedIn) {
+        this.api.updateDatasetView(this.datasetView);
+      }
     }
   }
 
@@ -1612,7 +1633,9 @@ export class Main extends VuexModule {
   }) {
     if (this.datasetView) {
       Vue.set(this.datasetView.layerContrasts, layerId, contrast);
-      this.api.updateDatasetView(this.datasetView);
+      if (this.isLoggedIn) {
+        this.api.updateDatasetView(this.datasetView);
+      }
     }
   }
 
@@ -1620,7 +1643,9 @@ export class Main extends VuexModule {
   async resetContrastInView(layerId: string) {
     if (this.datasetView) {
       Vue.delete(this.datasetView.layerContrasts, layerId);
-      this.api.updateDatasetView(this.datasetView);
+      if (this.isLoggedIn) {
+        this.api.updateDatasetView(this.datasetView);
+      }
     }
   }
 
@@ -1648,7 +1673,9 @@ export class Main extends VuexModule {
   }) {
     if (this.datasetView) {
       Vue.set(this.datasetView.scales, itemId, scale);
-      this.api.updateDatasetView(this.datasetView);
+      if (this.isLoggedIn) {
+        this.api.updateDatasetView(this.datasetView);
+      }
     }
   }
 
@@ -1656,7 +1683,9 @@ export class Main extends VuexModule {
   resetScalesInView(itemId: keyof IScales) {
     if (this.datasetView) {
       Vue.delete(this.datasetView.scales, itemId);
-      this.api.updateDatasetView(this.datasetView);
+      if (this.isLoggedIn) {
+        this.api.updateDatasetView(this.datasetView);
+      }
     }
   }
 
@@ -2025,7 +2054,9 @@ export class Main extends VuexModule {
       return;
     }
     Vue.set(this.datasetView, "layerContrasts", {});
-    this.api.updateDatasetView(this.datasetView);
+    if (this.isLoggedIn) {
+      this.api.updateDatasetView(this.datasetView);
+    }
   }
 
   @Mutation
@@ -2034,7 +2065,9 @@ export class Main extends VuexModule {
       return;
     }
     Vue.set(this.datasetView, "layerContrasts", contrasts);
-    this.api.updateDatasetView(this.datasetView);
+    if (this.isLoggedIn) {
+      this.api.updateDatasetView(this.datasetView);
+    }
   }
 
   @Action
@@ -2102,7 +2135,9 @@ export class Main extends VuexModule {
     [key: string]: string;
   }): Promise<void> {
     try {
-      await this.api.setUserColors(channelColors);
+      if (this.isLoggedIn) {
+        await this.api.setUserColors(channelColors);
+      }
 
       // Update the user metadata in the store using mutation
       this.updateUserChannelColors(channelColors);
